--- a/src/sif/hierarchylimits.cc
+++ b/src/sif/hierarchylimits.cc
@@ -1,0 +1,7 @@
+#include "sif/hierarchylimits.h"
+
+using namespace valhalla::sif;
+
+bool HierarchyLimits::StopExpanding(const float dist) const {
+  return (dist > expansion_within_dist && up_transition_count > max_up_transitions);
+}