--- a/src/thor/timedep_forward.cc
+++ b/src/thor/timedep_forward.cc
@@ -15,7 +15,7 @@ namespace thor {
 constexpr uint64_t kInitialEdgeLabelCount = 500000;
 
 // Number of iterations to allow with no convergence to the destination
-constexpr uint32_t kMaxIterationsWithoutConvergence = 200000;
+constexpr uint32_t kMaxIterationsWithoutConvergence = 800000;
 
 // Default constructor
 TimeDepForward::TimeDepForward() : AStarPathAlgorithm() {