--- a/src/thor/worker.cc
+++ b/src/thor/worker.cc
@@ -356,6 +356,8 @@ void thor_worker_t::parse_filter_attributes(const Api& request, bool is_strict_f
 void thor_worker_t::cleanup() {
   astar.Clear();
   bidir_astar.Clear();
+  timedep_forward.Clear();
+  timedep_reverse.Clear();
   multi_modal_astar.Clear();
   trace.clear();
   isochrone_gen.Clear();