--- a/valhalla/sif/hierarchylimits.h
+++ b/valhalla/sif/hierarchylimits.h
@@ -65,9 +65,7 @@ struct HierarchyLimits {
    * @param  dist  Distance (meters) from the destination.
    * @return  Returns true if expansion at this hierarchy level should stop.
    */
-  bool StopExpanding(const float dist) const {
-    return (dist > expansion_within_dist && up_transition_count > max_up_transitions);
-  }
+  bool StopExpanding(const float dist) const;
 
   /**
    * Determine if expansion of a hierarchy level should be stopped once