# Issue

We have a set of problematic routes on the Autobahn which end with
a rest-stop on the opposite side of travel direction, thereby
requiring overshooting the destination and coming back in the
other direction.

The bidirectional a-star routes fine but the forward timedependant requires
more iterations without convergence for these routes.

## Benchmarking

We're expecting complex routes to compute for a longer time, but to actually find a route. We want to ensure that average routes don't take longer to compute. Using the `de_benchmark_routes`, this change does not seem to affect regular routes.

Thor wall-clock times for `test_requests/de_benchmark_routes.txt`
```
master
Logfile de_benchmark_routes-kMaxIterations-200000.log contains 1714 elapsed times
  p10:    263ms
  mean:   444ms
  median: 415ms
  p90:    661ms
  max:    1675ms

This PR
Logfile de_benchmark_routes-kMaxIterations-800000.log contains 1731 elapsed times
  p10:    258ms
  mean:   439ms
  median: 411ms
  p90:    650ms
  max:    1616ms
```

## Tasklist
 - [x] Verify if change is necessary in reverse a-star as well
 - [ ] ~Add tests~
 - [x] Review - you must request approval to merge any PR to master
 - [x] Add #fixes with the issue number that this PR addresses
 - [x] Generally use squash merge to rebase and clean comments before merging
 - [x] Update the [changelog](CHANGELOG.md)
 - [x] Update relevant [documentation](docs/README.md)

## Requirements / Relations

 Link any requirements here. Other pull requests this PR is based on?